Change the winner to be a double to allow for scoring to have an amplitude to how much the player won by.

+++ repos/Algorithms/Solutions/Algorithms/Minimax.cs
@@ -37,7 +37,7 @@ namespace Algorithms
                 double eval = currentNode.Evaluate();
                 return eval;
             }
-            int? winner = currentNode.GetWinner();
+            double? winner = currentNode.GetWinner();
             if (winner.HasValue)
             {
                 return winner.Value;
